showNftEnablementToast
