+++ ui/selectors/selectors.js
@@ -771,10 +771,6 @@ export function getNftIsStillFetchingIndication(state) {
   return state.appState.isNftStillFetchingIndication;
 }
 
-export function getNftDetectionEnablementToast(state) {
-  return state.appState.showNftDetectionEnablementToast;
-}
-
 export function getCurrentCurrency(state) {
   return state.metamask.currentCurrency;
 }
